index: make build_entry raise on FileNotFound
Pre-requisite for fixing https://github.com/iterative/dvc/issues/9541
